#55 Changing metadata reader to be an interface
